Split sphinx extensions up a little
